added documentation

+++ coupling/ResultExporter/JSONExporter.cpp
@@ -24,7 +24,7 @@ void JSONExporter::exportCoupling(const CouplingGraph& couplingGraph)
     auto nodeList = couplingGraph.getNodes();
     std::map<std::string, int> nodeToIndexMap;
 
-    for (int i = 0; i < nodeList.size(); i++)
+    for (size_t i = 0; i < nodeList.size(); i++)
     {
         auto currentNode = nodeList[i];
         nodeToIndexMap[currentNode] = i;
